# Speed up content checks by not repeating requests