--- a/spec/features/content_pages_spec.rb
+++ b/spec/features/content_pages_spec.rb
@@ -24,8 +24,6 @@ RSpec.feature "content pages check", type: :feature, content: true do
       .and_return([])
   end
 
-  let(:document) { Nokogiri.parse(page.body) }
-
   # rubocop:disable RSpec/BeforeAfterAll
   before(:all) do
     statuses_deemed_successful = Rack::Utils::SYMBOL_TO_STATUS_CODE.values_at(:ok, :moved_permanently)
@@ -125,35 +123,36 @@ RSpec.feature "content pages check", type: :feature, content: true do
       end
     end
   end
-end
 
-describe "navbar" do
-  subject { page }
+  describe "navbar" do
+    subject { page }
 
-  before { visit "/" }
+    before { visit "/" }
+    let(:document) { Nokogiri.parse(page.body) }
 
-  let(:navigation_pages) { Pages::Frontmatter.select(:navigation) }
+    let(:navigation_pages) { Pages::Frontmatter.select(:navigation) }
 
-  scenario "navigable pages appear in desktop navbar" do
-    navigation_pages.each do |url, frontmatter|
-      page.within "nav .navbar__desktop" do
-        is_expected.to have_link frontmatter[:title], href: url
+    scenario "navigable pages appear in desktop navbar" do
+      navigation_pages.each do |url, frontmatter|
+        page.within "nav .navbar__desktop" do
+          is_expected.to have_link frontmatter[:title], href: url
+        end
       end
     end
-  end
 
-  scenario "navigable pages appear in mobile navbar" do
-    navigation_pages.each do |url, frontmatter|
-      page.within "nav .navbar__mobile" do
-        is_expected.to have_link frontmatter[:title], href: url
+    scenario "navigable pages appear in mobile navbar" do
+      navigation_pages.each do |url, frontmatter|
+        page.within "nav .navbar__mobile" do
+          is_expected.to have_link frontmatter[:title], href: url
+        end
       end
     end
-  end
 
-  scenario "mobile nav matches desktop nav" do
-    document.css("nav .navbar__desktop a").each do |desktop_link|
-      page.within "nav .navbar__mobile" do
-        is_expected.to have_link desktop_link.text, href: desktop_link["href"]
+    scenario "mobile nav matches desktop nav" do
+      document.css("nav .navbar__desktop a").each do |desktop_link|
+        page.within "nav .navbar__mobile" do
+          is_expected.to have_link desktop_link.text, href: desktop_link["href"]
+        end
       end
     end
   end