Previously every single page and link check used `visit` to load the page, this creates lots of duplicate (identical) requests.

Now the loops are run _inside_ the `scenario`, so it looks like there are fewer tests but there should be an equal number of assertions.

Some benchmarks:

```
 .../get-into-teaching-app speed-up-content-checks[8e217dd8] ✔ $
 ❯ time bundle exec rspec --tag content
Run options:
  include {:content=>true}
  exclude {:onschedule=>true}

Randomized with seed 48196
.............................................................................................................................................................................................................................................................................................................................................................................................

Finished in 3 minutes 51.4 seconds (files took 1.83 seconds to load)
381 examples, 0 failures

Randomized with seed 48196

bundle exec rspec --tag content  213.33s user 18.82s system 99% cpu 3:53.74 total

 .../get-into-teaching-app master[5eaeee71] ✔ $
 ❯ git switch speed-up-content-checks
Switched to branch 'speed-up-content-checks'

 .../get-into-teaching-app speed-up-content-checks[8e217dd8] ✔ $
 ❯ time bundle exec rspec --tag content
Use the new slots API (https://viewcomponent.org/guide/slots.html) instead. (called from <class:AccordionComponent> at /home/peter/projects/department-for-education/current/get-into-teaching-app/app/components/content/accordion_component.rb:6)
Run options:
  include {:content=>true}
  exclude {:onschedule=>true}

Randomized with seed 49356
......

Finished in 11.86 seconds (files took 1.83 seconds to load)
6 examples, 0 failures

Randomized with seed 49356

bundle exec rspec --tag content  13.02s user 1.06s system 99% cpu 14.160 total
```

From 213 seconds to 13 seconds. Not bad.